Major Refactors and Fixes, ToDo - Docs, Tests
