+++ keras_nlp/layers/__init__.py
@@ -16,7 +16,7 @@ from keras_nlp.layers.fnet_encoder import FNetEncoder
 from keras_nlp.layers.mlm_head import MLMHead
 from keras_nlp.layers.mlm_mask_generator import MLMMaskGenerator
 from keras_nlp.layers.position_embedding import PositionEmbedding
-from keras_nlp.layers.random_deletion import RandomDeletion
+from keras_nlp.layers.random_word_deletion import RandomWordDeletion
 from keras_nlp.layers.sine_position_encoding import SinePositionEncoding
 from keras_nlp.layers.token_and_position_embedding import (
     TokenAndPositionEmbedding,
